refactor: split pr9814 oversized files
